requirements.txt
fastapi
uvicorn[standard]
pydantic
sentence-transformers
faiss-cpu
transformers
torch
python-multipart
pypdf
python-jose
python-dotenv
sqlalchemy
requests
